refactor: Split aggregation.rs into focused modules
Split the 550-line aggregation.rs file into a modular structure:

- aggregation/mod.rs (~130 lines) - Main execution logic with execute_with_aggregation
- aggregation/detection.rs (~36 lines) - Aggregate detection helpers (has_aggregates, expression_has_aggregate)
- aggregation/evaluation.rs (~400 lines) - Expression evaluation logic for aggregates

All functions maintain their original pub(in crate::select::executor) visibility to ensure proper access from sibling modules in the executor package.

This refactor improves code organization and maintainability without changing any functionality. All existing tests pass.

Closes #432
