+++ crates/executor/src/select/executor/aggregation/evaluation.rs
@@ -1,168 +1,14 @@
-//! Aggregation execution methods for SelectExecutor
+//! Expression evaluation with aggregates for SelectExecutor
 
-use super::builder::SelectExecutor;
+use super::super::builder::SelectExecutor;
 use crate::errors::ExecutorError;
 use crate::evaluator::{CombinedExpressionEvaluator, ExpressionEvaluator};
-use crate::select::cte::CteResult;
-use crate::select::filter::apply_where_filter_combined;
-use crate::select::grouping::{group_rows, AggregateAccumulator};
-use crate::select::helpers::{apply_distinct, apply_limit_offset};
-use std::collections::HashMap;
+use crate::select::grouping::AggregateAccumulator;
 
 impl<'a> SelectExecutor<'a> {
-    /// Check if SELECT list contains aggregate functions
-    pub(super) fn has_aggregates(&self, select_list: &[ast::SelectItem]) -> bool {
-        select_list.iter().any(|item| match item {
-            ast::SelectItem::Expression { expr, .. } => self.expression_has_aggregate(expr),
-            _ => false,
-        })
-    }
-
-    /// Check if an expression contains aggregate functions
-    #[allow(clippy::only_used_in_recursion)]
-    pub(super) fn expression_has_aggregate(&self, expr: &ast::Expression) -> bool {
-        match expr {
-            // New AggregateFunction variant
-            ast::Expression::AggregateFunction { .. } => true,
-            // Old Function variant (backwards compatibility)
-            ast::Expression::Function { name, .. } => {
-                matches!(name.to_uppercase().as_str(), "COUNT" | "SUM" | "AVG" | "MIN" | "MAX")
-            }
-            ast::Expression::BinaryOp { left, right, .. } => {
-                self.expression_has_aggregate(left) || self.expression_has_aggregate(right)
-            }
-            _ => false,
-        }
-    }
-
-    /// Execute SELECT with aggregation/GROUP BY
-    pub(super) fn execute_with_aggregation(
-        &self,
-        stmt: &ast::SelectStmt,
-        cte_results: &HashMap<String, CteResult>,
-    ) -> Result<Vec<storage::Row>, ExecutorError> {
-        // Execute FROM clause (handles JOINs, subqueries, CTEs)
-        let from_result = match &stmt.from {
-            Some(from_clause) => self.execute_from(from_clause, cte_results)?,
-            None => {
-                return Err(ExecutorError::UnsupportedFeature(
-                    "SELECT without FROM not yet implemented".to_string(),
-                ))
-            }
-        };
-
-        eprintln!(
-            "DEBUG AGG: schema keys={:?}",
-            from_result.schema.table_schemas.keys().collect::<Vec<_>>()
-        );
-        eprintln!(
-            "DEBUG AGG: outer_row={}, outer_schema={:?}",
-            self._outer_row.is_some(),
-            self._outer_schema.map(|s| s.table_schemas.keys().collect::<Vec<_>>())
-        );
-
-        // Create evaluator with outer context if available (outer schema is already a CombinedSchema)
-        let evaluator =
-            if let (Some(outer_row), Some(outer_schema)) = (self._outer_row, self._outer_schema) {
-                eprintln!(
-                    "DEBUG AGG: Creating evaluator WITH outer context, outer tables={:?}",
-                    outer_schema.table_schemas.keys().collect::<Vec<_>>()
-                );
-                CombinedExpressionEvaluator::with_database_and_outer_context(
-                    &from_result.schema,
-                    self.database,
-                    outer_row,
-                    outer_schema,
-                )
-            } else {
-                eprintln!("DEBUG AGG: Creating evaluator WITHOUT outer context");
-                CombinedExpressionEvaluator::with_database(&from_result.schema, self.database)
-            };
-
-        // Apply WHERE clause to filter joined rows
-        let filtered_rows =
-            apply_where_filter_combined(from_result.rows, stmt.where_clause.as_ref(), &evaluator)?;
-
-        // Group rows
-        let groups = if let Some(group_by_exprs) = &stmt.group_by {
-            group_rows(&filtered_rows, group_by_exprs, &evaluator)?
-        } else {
-            // No GROUP BY - treat all rows as one group
-            vec![(Vec::new(), filtered_rows)]
-        };
-
-        // Compute aggregates for each group and apply HAVING
-        let mut result_rows = Vec::new();
-        for (group_key, group_rows) in groups {
-            // Compute aggregates for this group
-            let mut aggregate_results = Vec::new();
-            for item in &stmt.select_list {
-                match item {
-                    ast::SelectItem::Expression { expr, .. } => {
-                        let value = self.evaluate_with_aggregates(
-                            expr,
-                            &group_rows,
-                            &group_key,
-                            &evaluator,
-                        )?;
-                        aggregate_results.push(value);
-                    }
-                    ast::SelectItem::Wildcard => {
-                        return Err(ExecutorError::UnsupportedFeature(
-                            "SELECT * not supported with aggregates".to_string(),
-                        ))
-                    }
-                }
-            }
-
-            // Apply HAVING filter
-            let include_group = if let Some(having_expr) = &stmt.having {
-                let having_result = self.evaluate_with_aggregates(
-                    having_expr,
-                    &group_rows,
-                    &group_key,
-                    &evaluator,
-                )?;
-                match having_result {
-                    types::SqlValue::Boolean(true) => true,
-                    types::SqlValue::Boolean(false) | types::SqlValue::Null => false,
-                    other => {
-                        return Err(ExecutorError::InvalidWhereClause(format!(
-                            "HAVING must evaluate to boolean, got: {:?}",
-                            other
-                        )))
-                    }
-                }
-            } else {
-                true
-            };
-
-            if include_group {
-                result_rows.push(storage::Row::new(aggregate_results));
-            }
-        }
-
-        // Apply ORDER BY if present
-        let result_rows = if let Some(order_by) = &stmt.order_by {
-            self.apply_order_by_to_aggregates(result_rows, stmt, order_by)?
-        } else {
-            result_rows
-        };
-
-        // Apply DISTINCT if specified
-        let result_rows = if stmt.distinct { apply_distinct(result_rows) } else { result_rows };
-
-        // Don't apply LIMIT/OFFSET if we have a set operation - it will be applied later
-        if stmt.set_operation.is_some() {
-            Ok(result_rows)
-        } else {
-            Ok(apply_limit_offset(result_rows, stmt.limit, stmt.offset))
-        }
-    }
-
     /// Evaluate an expression in the context of aggregation
     #[allow(clippy::only_used_in_recursion)]
-    pub(super) fn evaluate_with_aggregates(
+    pub(in crate::select::executor) fn evaluate_with_aggregates(
         &self,
         expr: &ast::Expression,
         group_rows: &[storage::Row],
@@ -227,7 +73,7 @@ impl<'a> SelectExecutor<'a> {
 
     /// Evaluate aggregate function expressions (COUNT, SUM, AVG, MIN, MAX)
     /// Handles both AggregateFunction and Function variants (for backwards compatibility)
-    fn evaluate_aggregate_function(
+    pub(in crate::select::executor) fn evaluate_aggregate_function(
         &self,
         expr: &ast::Expression,
         group_rows: &[storage::Row],
@@ -281,7 +127,7 @@ impl<'a> SelectExecutor<'a> {
     }
 
     /// Evaluate binary operations in aggregate context
-    fn evaluate_binary_op_with_aggregates(
+    pub(in crate::select::executor) fn evaluate_binary_op_with_aggregates(
         &self,
         left: &ast::Expression,
         op: &ast::BinaryOperator,
@@ -300,7 +146,7 @@ impl<'a> SelectExecutor<'a> {
     }
 
     /// Evaluate scalar subqueries and EXISTS expressions in aggregate context
-    fn evaluate_scalar_subquery_with_aggregates(
+    pub(in crate::select::executor) fn evaluate_scalar_subquery_with_aggregates(
         &self,
         expr: &ast::Expression,
         group_rows: &[storage::Row],
@@ -315,7 +161,7 @@ impl<'a> SelectExecutor<'a> {
     }
 
     /// Evaluate IN predicate with subquery in aggregate context
-    fn evaluate_in_predicate_with_aggregates(
+    pub(in crate::select::executor) fn evaluate_in_predicate_with_aggregates(
         &self,
         left_expr: &ast::Expression,
         subquery: &ast::SelectStmt,
@@ -374,7 +220,7 @@ impl<'a> SelectExecutor<'a> {
     }
 
     /// Evaluate quantified comparison (ALL/ANY/SOME) with subquery in aggregate context
-    fn evaluate_quantified_comparison_with_aggregates(
+    pub(in crate::select::executor) fn evaluate_quantified_comparison_with_aggregates(
         &self,
         left_expr: &ast::Expression,
         op: &ast::BinaryOperator,
@@ -483,7 +329,7 @@ impl<'a> SelectExecutor<'a> {
     }
 
     /// Apply ORDER BY to aggregated results
-    fn apply_order_by_to_aggregates(
+    pub(in crate::select::executor) fn apply_order_by_to_aggregates(
         &self,
         rows: Vec<storage::Row>,
         stmt: &ast::SelectStmt,
